requirements.txt
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.14.0
temporalio>=1.5.0
aiomysql>=0.2.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit-antd-components>=0.3.0
streamlit-extras>=0.3.0
streamlit-float>=0.3.0
google-generativeai>=0.3.0
pandas>=2.0.0
plotly>=5.15.0
requests>=2.31.0
